requirements.txt
numpy
pandas
matplotlib
python-dotenv
openai>=1.0.0
requests 